Extracted `MergedXml` interface
